+++ quantforge/__init__.py
@@ -774,6 +774,7 @@ from .matrix_pencil import matrix_pencil
 from .music import music_pseudospectrum, music_frequencies
 from .randomized_svd import randomized_svd
 from .nmf import nmf
+from .dmd import dmd
 from .dual import Dual, derivative as dual_derivative
 from .dual_calculus import dual_gradient, dual_newton
 from .hyperdual import HyperDual, derivatives as hyperdual_derivatives, second_derivative
@@ -1478,7 +1479,7 @@ try:  # pragma: no cover - trivial availability branch
 except ImportError:  # pragma: no cover
     HAS_NUMPY = False
 
-__version__ = "5.62.0"
+__version__ = "5.63.0"
 
 __all__ = [
     "OptionType",
@@ -2521,6 +2522,7 @@ __all__ = [
     "music_frequencies",
     "randomized_svd",
     "nmf",
+    "dmd",
     "Dual",
     "dual_derivative",
     "dual_gradient",